requirements-dev.txt
# DO NOT EDIT THIS FILE DIRECTLY - use ./script/update-requirements to update
Pygments==2.19.1
SecretStorage==3.3.3
black==24.10.0
build==1.2.2.post1
cffi==1.17.1
click==8.1.8
cmarkgfm==2024.11.20
coverage==7.6.12
cryptography==44.0.1
docutils==0.21.2
id==1.5.0
iniconfig==2.0.0
isort==6.0.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jeepney==0.8.0
keyring==25.6.0
markdown-it-py==3.0.0
mdurl==0.1.2
more-itertools==10.6.0
mypy-extensions==1.0.0
nh3==0.2.20
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
pluggy==1.5.0
pycparser==2.22
pyflakes==3.2.0
pyproject_hooks==1.2.0
pytest-cov==6.0.0
pytest==8.3.4
pytest_network==0.0.1
readme_renderer==44.0
requests-mock==1.12.1
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.9.4
twine==6.1.0
